Refactor AWS S3
